requirements.txt
altgraph==0.17
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
dataclasses==0.8
filelock==3.0.12
future==0.18.2
huggingface-hub==0.0.8
idna==2.10
importlib-metadata==4.5.0
joblib==1.0.1
numpy==1.19.5
packaging==20.9
pefile==2021.5.24
pycparser==2.20
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
pyparsing==2.4.7
pywin32-ctypes==0.2.0
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
sacremoses==0.0.45
six==1.16.0
SoundFile==0.10.3.post1
tokenizers==0.10.3
torch==1.9.0
tqdm==4.61.1
transformers==4.7.0
typing-extensions==3.10.0.0
urllib3==1.26.5
zipp==3.4.1
scipy
librosa==0.8.0
scikit-learn==0.23.2
websockets
ffmpeg-python==0.2.0